requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
# SQLAlchemy 2.0.23 has issues on Python 3.13; bump to a 2.0.x version with 3.13 support
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
stripe==7.8.0
jinja2==3.1.2
python-dotenv==1.0.0
psycopg2-binary
alembic
slowapi==0.1.9
aiofiles==23.2.1
bleach==6.1.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
pytest-cov==4.1.0